Er ikkje så glad i util-filer, dei har det med å trekkje til seg metodar og logikk som burde liggje andre stadar. Så nyttar høvet til å flytte hentPerioderMedUtbetaling-metoden ut derifrå og slette heile util-klassa. Litt usikker på kor denne helst bør liggje, men startar med å flytte han til service-fila, det er jo derifrå han brukast per no
